refs CRASM-3382: Restructure component tests to mirror actual component organization
- header.test.tsx -> Header/header.test.tsx
- routeGuard.test.tsx -> Routes/routeGuard.test.tsx
- updateUserStateForm.test.tsx -> UpdateUserStateForm/updateUserStateForm.test.tsx
- termsOfUse.test.tsx -> Dialog/TermsOfUse/termsOfUse.test.tsx
- vs-dashboard-gate.test.tsx -> Gates/vs-dashboard-gate.test.tsx
- Update import paths to reflect new directory structure
- Root-level components (govBanner.test.tsx, layout.test.tsx) remain at components root.
- All moved tests verified working with correct snapshots generated.
