+++ frontend/src/tests/components/UpdateUserStateForm/updateUserStateForm.test.tsx
@@ -4,10 +4,10 @@ import {
   fireEvent,
   waitFor,
   screen
-} from '../../test-utils/test-utils';
-import { testUser } from '../../test-utils/user';
+} from '../../../test-utils/test-utils';
+import { testUser } from '../../../test-utils/user';
 import { describe, expect, it, vi, beforeEach } from 'vitest';
-import { UpdateStateForm } from '../../components/UpdateUserStateForm/UpdateUserStateForm';
+import { UpdateStateForm } from '../../../components/UpdateUserStateForm/UpdateUserStateForm';
 
 vi.mock('@/constants/constants', () => ({
   STATE_OPTIONS: ['California', 'Texas', 'New York', 'Florida']
